fix: add WebGPUNativeType
